pyproject.toml
[project]
name = "capital-com-test-task"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "flask>=3.1.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2>=2.9.10",
    "pydantic>=2.10.6",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
]


[tool.black]
line-length = 79
exclude = '''
/(
  | migrations
  | .venv
)/
'''